# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Absensi Mahasiswa Geolocation API",
        "description": "API untuk sistem absensi mahasiswa berbasis geolocation. Mahasiswa dapat melakukan absensi dengan validasi lokasi GPS.",
        "contact": {
            "name": "Admin Support",
            "email": "admin@kampus.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        }
    ],
    "paths": {
...
        "/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Daftar semua jadwal",
                "description": "Mendapatkan semua jadwal dengan detail course dan location (Admin only)",
                "operationId": "indexSchedules",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "course_id": {
                                                        "type": "integer"
                                                    },
                                                    "location_id": {
                                                        "type": "integer"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "course": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "course_name": {
                                                                "type": "string"
                                                            },
                                                            "course_code": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "location": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "latitude": {
                                                                "type": "number"
                                                            },
                                                            "longitude": {
                                                                "type": "number"
                                                            },
                                                            "radius": {
                                                                "type": "number"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Tambah jadwal baru",
                "description": "Membuat jadwal baru (Admin only)",
                "operationId": "storeSchedule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "location_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-15 08:00:00"
                                    },
                                    "end_time": {
                                        "description": "Harus setelah start_time",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-15 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}